MAINT: Move collab_pls to an nd pls mixin method
