+++ pybaselines/two_d/optimizers.py
@@ -16,125 +16,15 @@ from math import ceil
 
 import numpy as np
 
+from .._nd.optimizers import _OptimizersNDMixin
 from .._validation import _check_optional_array, _get_row_col_values
 from ..api import Baseline
 from ..utils import _check_scalar, _sort_array2d
 from ._algorithm_setup import _Algorithm2D
 
 
-class _Optimizers(_Algorithm2D):
+class _Optimizers(_Algorithm2D, _OptimizersNDMixin):
     """A base class for all optimizer algorithms."""
-
-    @_Algorithm2D._handle_io(ensure_dims=False, skip_sorting=True)
-    def collab_pls(self, data, average_dataset=True, method='asls', method_kwargs=None):
-        """
-        Collaborative Penalized Least Squares (collab-PLS).
-
-        Averages the data or the fit weights for an entire dataset to get more
-        optimal results. Uses any Whittaker-smoothing-based or weighted spline algorithm.
-
-        Parameters
-        ----------
-        data : array-like, shape (L, M, N)
-            An array with shape (L, M, N) where L is the number of entries in
-            the dataset and (M, N) is the shape of each data entry.
-        average_dataset : bool, optional
-            If True (default) will average the dataset before fitting to get the
-            weighting. If False, will fit each individual entry in the dataset and
-            then average the weights to get the weighting for the dataset.
-        method : str, optional
-            A string indicating the Whittaker-smoothing-based or weighted spline method to
-            use for fitting the baseline. Default is 'asls'.
-        method_kwargs : dict, optional
-            A dictionary of keyword arguments to pass to the selected `method` function.
-            Default is None, which will use an empty dictionary.
-
-        Returns
-        -------
-        baselines : np.ndarray, shape (L, M, N)
-            An array of all of the baselines.
-        params : dict
-            A dictionary with the following items:
-
-            * 'average_weights': numpy.ndarray, shape (M, N)
-                The weight array used to fit all of the baselines.
-            * 'average_alpha': numpy.ndarray, shape (M, N)
-                Only returned if `method` is 'aspls'. The
-                `alpha` array used to fit all of the baselines for the
-                :meth:`~.Baseline2D.aspls`.
-            * 'method_params': dict[str, list]
-                A dictionary containing the output parameters for each individual fit.
-                Keys will depend on the selected method and will have a list of values,
-                with each item corresponding to a fit.
-
-        Raises
-        ------
-        ValueError
-            Raised if the input data is not three dimensional.
-
-        Notes
-        -----
-        If `method` is 'aspls', `collab_pls` will also calculate
-        the `alpha` array for the entire dataset in the same manner as the weights.
-
-        References
-        ----------
-        Chen, L., et al. Collaborative Penalized Least Squares for Background
-        Correction of Multiple Raman Spectra. Journal of Analytical Methods
-        in Chemistry, 2018, 2018.
-
-        """
-        dataset, optimizer_obj, method_kws = self._setup_optimizer(
-            data, method, method_param={None: 'weights'}, method_kwargs=method_kwargs,
-            copy_kwargs=True
-        )
-        data_shape = dataset.shape
-        if len(data_shape) != 3:
-            raise ValueError((
-                'the input data must have a shape of (number of measurements, number of x points,'
-                f' number of y points), but instead has a shape of {data_shape}'
-            ))
-        # if using aspls or pspline_aspls, also need to calculate the alpha array
-        # for the entire dataset
-        calc_alpha = optimizer_obj.method in ('aspls', 'pspline_aspls')
-
-        # step 1: calculate weights for the entire dataset
-        if average_dataset:
-            _, fit_params = optimizer_obj.method_call(np.mean(dataset, axis=0), **method_kws)
-            method_kws['weights'] = fit_params['weights']
-            if calc_alpha:
-                method_kws['alpha'] = fit_params['alpha']
-        else:
-            weights = np.empty(data_shape)
-            if calc_alpha:
-                alpha = np.empty(data_shape)
-            for i, entry in enumerate(dataset):
-                _, fit_params = optimizer_obj.method_call(entry, **method_kws)
-                weights[i] = fit_params['weights']
-                if calc_alpha:
-                    alpha[i] = fit_params['alpha']
-            method_kws['weights'] = np.mean(weights, axis=0)
-            if calc_alpha:
-                method_kws['alpha'] = np.mean(alpha, axis=0)
-
-        # step 2: use the dataset weights from step 1 (stored in method_kws['weights'])
-        # to fit each individual data entry; set tol to infinity so that only one
-        # iteration is done and new weights are not calculated
-        if 'tol' in optimizer_obj.method_signature.parameters:
-            method_kws['tol'] = np.inf
-        if 'tol_2' in optimizer_obj.method_signature.parameters:  # brpls
-            method_kws['tol_2'] = np.inf
-        baselines = np.empty(data_shape)
-        params = {'average_weights': method_kws['weights'], 'method_params': defaultdict(list)}
-        if calc_alpha:
-            params['average_alpha'] = method_kws['alpha']
-
-        for i, entry in enumerate(dataset):
-            baselines[i], param = optimizer_obj.method_call(entry, **method_kws)
-            for key, value in param.items():
-                params['method_params'][key].append(value)
-
-        return baselines, params
 
     @_Algorithm2D._handle_io(skip_sorting=True)
     def adaptive_minmax(self, data, poly_order=None, method='modpoly', weights=None,
